Migrate Season Catalog & Season Card to Typescript
