--- src/components/season/SeasonCard/SeasonCard.styles.js
@@ -1,5 +1,0 @@
-import styled from 'styled-components';
-import { Link } from 'gatsby';
-
-export const SeasonCardLink = styled(Link)``;
-export const SeasonCard = styled.article``;
