--- src/components/season/SeasonCatalog/SeasonCatalog.styles.js
@@ -1,6 +1,0 @@
-import styled from 'styled-components';
-
-export const SeasonCatalog = styled.div`
-	max-width: 1200px;
-	margin: 0 auto;
-`;
